Updated LangChain & Requirements
I was getting a failed build on Python 3.8 because the packages have been updated and no longer work with Python 3.8. The right thing to do would be to fix package numbers but I couldn't get the right combination myself. This works on 3.10 and 3.11 and adopts the new langchain code changes and supports stable.